create standalone js file for RCTPicker
